Phase 4 Step 5 누락 설정 보완 및 영속화 Version 타입을 long으로 변경
- 누락된 PlayerStateEntityConfiguration을 추가해 JSONB, 기본값, 동시성 토큰, cascade 및 unique index 설정 반영
- Inventory, Equipment, PlayerState의 Version을 int에서 long으로 변경
- persistence state, snapshot, commit 및 EF Core 동시성 처리의 Version 타입을 long으로 통일
- PostgreSQL version 컬럼을 integer에서 bigint로 변경하는 migration 및 model snapshot 추가
- PlayerCacheService의 캐시 키 생성을 PlayerPersistenceCacheKeys로 통일
- PlayerCacheService에 남아 있던 직접 DB 저장 및 캐시 갱신 경로를 제거해 쓰기 소유권을 PlayerPersistenceService로 일원화

+++ Server/DatabaseLib/Entities/EquipmentEntity.cs
@@ -19,7 +19,7 @@ namespace DatabaseLib.Entities
 		public long PlayerId { get; set; }
 
 		[Column( "version" )]
-		public int Version { get; set; } = 1; // 낙관적 동시성 제어용?
+		public long Version { get; set; } = 1; // 낙관적 동시성 제어용?
 
 		[Column( "last_updated" )]
 		public DateTime LastUpdated { get; set; } = DateTime.UtcNow;
