+++ Server/Server/Services/Persistence/PlayerSaveCommit.cs
@@ -6,8 +6,8 @@ using System.Threading.Tasks;
 
 namespace Server.Services.Persistence
 {
-	public sealed record PlayerSaveCommit( long PlayerRawId, long SavedRevision, int InventoryVersion,
-		int EquipmentVersion, int PlayerStateVersion );
+	public sealed record PlayerSaveCommit( long PlayerRawId, long SavedRevision, long InventoryVersion,
+		long EquipmentVersion, long PlayerStateVersion );
 
 	public sealed record CreatePlayerCommit( long PlayerRawId, long InventoryId, long EquipmentId, long PlayerStateId );
 }
